backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
supabase
websockets
python-dotenv==1.0.1
python-multipart==0.0.9
pydantic==2.6.1
